Create Derangement problem.ipynb
A derangement of the numbers 1 through n is a permutation of all n those numbers such that none of them is in the “right place.” For example, 34251 is a derangement of 1 through 5, but 24351 is not because 3 is in the 3rd position.We will use simulation to estimate the number of derangements of the numbers 1 through 12.

Here is a program that generates random permutations of the integers 1, 2,…, 12. The program determines whether or not each permutation is a derangement.